clean up the repo
